move table name tests to core.rs
